app/backend/requirements.txt
azure-identity==1.23.1
dotenv==0.9.9
fastapi
pyodbc
sqlalchemy
uvicorn
